backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
install==1.3.5
Pillow==9.3.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
django4-background-tasks
